Import the bootstrap into the sdk using a 'junction' element
This depends on the new 'junction' element that has recently been
merged into BuildStream's master branch.

Instead of having to build the bootstrap, check out the results to
fixed locations manually and then build the SDK, it should now be
possible to run `bst build` inside the SDK project and automatically
build or pull the correct bootstrap.

The bootstrap can be cross-compiled in some cases, using the new
`bootstrap_target_arch` option added in the SDK project. The sdk has
no support for cross-compilation, so I expect cross building will only
work on compatible architectures:

For example this would work for building an i586 SDK on x86_64:

  bootstrap_build_arch=x86_64 bootstrap_target_arch=i586 target_arch=i586

This would not be possible however as the x86_64 machine would not be
able to execute the aarch64 binaries:

  bootstrap_build_arch=x86_64 bootstrap_target_arch=aarch64 target_arch=aarch64