Add initial support for building bootable VM images
The primary purpose of the Freedesktop SDK is to power application
containers, but with the addition of a bootloader and kernel we can
also use it to boot virtual machines.

The VM images are intended for testing purposes only -- it's not a
goal to produce something aimed at end-users. However it can be very
useful for developers to test their changes to system components in
a fully integrated VM image, and this approach is more reproducible
and reliable than testing system components in a manually hacked-up
development environment. This mechanism could also serve as a basis
for automated testing like for example GNOME Continuous.

There is currently no init system in the VM image, instead it boots
directly into a shell.

Note that the x86image plugin can be made twice as fast with this patch:
https://gitlab.com/BuildStream/bst-external/merge_requests/16