python: Nuke tkinter out
The tkinter package (“Tk interface”) is the standard Python interface
to the Tk GUI toolkit.

At least that's what the upstream Python docs say, but nobody really use
Tk any more for GUI apps these days anyway.

The turtle module is a demo, built with tkinter, of the famous interface
to teach programming to kids. (by controlling a Turtle)

In addition, Tk is not even in the Platform/Sdk, so these modules won't
even work properly.

This saves about 5MB from a Platform/Sdk checkout.