base/tzdata: Remove the make arguments
The element was installing a few files in the wrong places:

    /usr/usr/bin/tzselect
    /usr/usr/bin/zdump
    /usr/usr/lib/libtz.a
    /usr/usr/sbin/zic

That's because they were installed respectively in:

*   BINDIR = $(TOPDIR)/$(USRDIR)/bin
*   LIBDIR = $(TOPDIR)/$(USRDIR)/lib
*   ZICDIR = $(TOPDIR)/$(USRDIR)/sbin

And USRDIR is defined as just "usr".

Since we were defining TOPDIR=/usr, we ended with the /usr/usr/… paths.

The other two make-args are now completely unnecessary though, they were
only there to workaround the wrong paths for the man pages and the
zoneinfo data. With an empty TOPDIR, those files are now installed in
the right place directly and we can stop defining MANDIR and TZDIR.

Relates to #202