bootstrap: Include gnu-config
Sometimes, apps bundle old versions of modules which were released
before a certain architecture they want to build on.

Very often, using the gnu-config's config.guess and/or config.sub is
enough to get the module to build on the newer arch.

Adding gnu-config to bootstrap will make it end up in the Sdk.

This was already done in org.freedesktop.Sdk//1.6 for the same reasons
(it is in Yocto).

Fixes #280