sdk/elements/desktop/ffmpeg.bst: correct license config
I had previously misunderstood the meaning/purpose of --enable-nonfree in
ffmpeg's configure script. By default ffmpeg is LGPL and hence permits linking
with GPL-incompatible libraries such as fdk-aac, but if you --enable-gpl and
transform ffmpeg into a GPL codebase by including GPL dependencies, this is no
longer the case. Passing --enable-nonfree ignores this conflict, but the result
is not distributable. However, in the SDK we definitely want libfdk-aac and do
not actually use any of the GPL dependencies of ffmpeg, so we can stop passing
--enable-gpl. This avoids the (hypothetical) incompatibility by building
ffmpeg as LGPL and means we do not need to pass --enable-nonfree any longer.