base/sqlite: Append to CPPFLAGS, don't redefine them
We currently completely redefine the CPPFLAGS passed to the build system
just to add a single option. This means we have to copy over the default
flags (from the project configuration) every time we update them, which
we will obviously forget eventually.

This commit instead just appends the new option, to ease future
maintenance.

Relates to #312