Patch gstreamer-libav to build
The easiest and cleanest (and presumably safest) way to patch
was to apply the fixes upstream had applied on their master branch,
however there were many fixes, and almost all commits between the
1.14.0 tag and the commit used here are API patches for ffmpeg, so
instead of using a tag and patching forward, I used the commit
where upstream declare a dependency on ffmpeg n4.0.x and remove
the one patch which uses GStreamer's unstable API. Thankfully this
doesn't seem to replace anything, so *shouldn't* break anything.

Additionally I expose gstreamer-libav to the SDK.

With the html5-codecs extension added, it is also necessary to stop
gst-libav caching the codecs installed, as this now may change
without the cache updating. This applies a (slightly) modified
version of this patch from debian [1].

[1]: https://salsa.debian.org/gstreamer-team/gst-libav1.0/blob/c4de9af16a673683f259ef440e3fcbd43d719eda/debian/patches/02_plugin-dependencies.patch