vaapi-intel: Fix the ref of the extended runtime
In a Flatpak repo, an OSTree ref is:

    $(type)/$(id)/$(arch)/$(branch)

where `type` is either "app" or "runtime".

Partial refs, like `$(id)/$(arch)/$(branch)` or even `$(id)` are often
used in the CLI as shortcuts, as Flatpak manages to resolve them when
there are no ambiguities.

When defining extensions though, we have to specify the full ref.

https://gitlab.gnome.org/GNOME/gnome-sdk-images/issues/7