Fix html5 codec build
When refactoring using includes[1], I neglected to change the prefix
on html5/ffmpeg, resulting in it being installed inside the SDK and
not at its mountpoint. As a result installing the extension did
the square root of nothing.

This changes the config to use variables for both prefix and libdir
of ffmpeg.

[1] https://gitlab.com/freedesktop-sdk/freedesktop-sdk/commit/6c6681854b6eeaacc036fb0b12c5b47d845053b2