desktop/libglvnd: Upgrade to the latest upstream release
A function was removed from the exported symbols with this new version,
which breaks ABI.

However, this should be entirely safe because the function was misnamed:
it was called glMultiDrawElementsBaseVertexOES when it should have been
called glMultiDrawElementsBaseVertexEXT.

And in fact, the latter also existed, and is still present.

No application should have been using this function, so let's ignore the
removal.