utils/check-abi: Ignore the private symbols when comparing ABI
This tells abidiff to only compare the ABI of types which can be found
in the installed headers.

It not only reduces false positives for ABI breaks (for example, we
don't get the ignored libgtk-3.so.0 ABI changes), but it should also
consume a bit less memory, which is always a good thing.

This requires using the Sdk though, as the Platform doesn't have the
development headers.