utils/check-abi: Simplify the way we match libraries
In the Platform, the only way we could find the libraries was to search
for all the lib*.so* files, and ensure they aren't debuginfo.

Now that we check the ABI of the Sdk, we have access to the development
symlinks.

As a result, we can simply look for all the lib*.so files.