desktop/libsoup: Upgrade to 2.64.2
With this release, it now builds with Meson, which is progress on #355.

However, we have to explicitly disable the gssapi integration. With the
Autotools, the build system would automatically disable it when the
dependency wasn't present, but Meson hard fails on it.

This now requires libpsl, which was just added in the previous commit.

While we're at it, let's use a Git tag instead of a tarball, and use the
new ref format. (#441)