base/glib: Upgrade to 2.58.1
With this release, it now builds with Meson, which is progress on #355.

The patch we had is now completely unnecessary.

While we're at it, let's use the URL alias, and use the new ref format.
(#441)

With Meson, the libselinux support is enabled by default, so we have to
disable it explicitly, as it was disabled with the Autotools build.

Also, we need to add a dependency on m4, so that the various macro files
are generated and installed. These are required for elements depending
on GLib but using the Autotools.