utils/check-abi: Ignore the gdk-pixbuf internals
There is no point in checking the ABI of the image loaders, as they
aren't libraries things will link against, they are only used internally
by gdk-pixbuf itself.

As for the main gdk-pixbuf library, Emmanuele Bassi says that the public
symbols are only the ones following a certain namespace, even though
others might be incorrectly exposed.

As it turns out, when upgrading gdk-pixbuf to 2.38.0 the ABI checker
reports ABI breaks on the above 2 reasons.