desktop/gtk3: Properly update the icon cache
The gtk-update-icon-cache command takes a path to an icon theme as an
argument.

Calling it without any argument does nothing, but returns 0, so we never
realized it wasn't updating any cache.

Of course, we can't update the hicolor icon theme cache unless the
hicolor theme is installed at the time we run the command, so we must
add an explicit dependency.