base: Add an OpenSSL 1.0 element
We currently have OpenSSL 1.0 as base/openssl.bst, but we really want to
update that to 1.1.

Unfortunately, OpenSSH doesn't build against OpenSSL 1.1 yet, so we need
some sort of compatibility.

This new base/openssl-1.0.bst element serves this purpose: when we
update openssl.bst to 1.1, OpenSSH can continue building against 1.0.

However, we do not want to expose both versions in the Sdk, for apps to
link against. First, because it creates a nightmare of overlapping files
(pkgconfig, headers, .so symlinks, …), and second because we want
everything to use modern crypto.

So openssl-1.0.bst can be a build dependency for those elements which
absolutely require it (they get the development files in order to
build), but they will have a runtime dependency on
openssl-1.0-runtime-only.bst, which excludes all those development
files.

At the boundary of the Sdk, OpenSSL 1.0 effectively becomes a
"runtime-only" library.