utils/check-abi: Check the ABI of the Platform
When we first introduced this script in !398, is was checking the ABI of
the Platform, as that's really the right thing to do.

Unfortunately it would report lots of false positives. Since we want to
block merging branches on ABI breaks, it is critical that we eliminate
as many false positives as possible.

In order to do so, we started telling abidiff to ignore the private
symbols in !498, as if those break ABI it won't affect apps.

However, that requires access to the installed, public library headers,
which only are in the Sdk, not in the Platform. And so we moved to
checking the ABI of the Sdk instead of the Platform at the same time.

This was not really a good idea, and when I made that change I knew it
wasn't ideal. However, I judged that in practice it wouldn't cause too
much trouble, since the Sdk and Platform aren't very different in terms
of the libraries they include.

Needless to say, reality proved me wrong. When we landed the OpenSSL
upgrade in !523, we forgot to add OpenSSL 1.0 to the Platform, which
broke ABI. (and most likely a lot of apps)

It ended up in the Sdk though, which means the ABI checker was happy,
and we didn't find the problem. Had we been checking the ABI of the
Platform, this issue would have been prevented because the CI would have
failed and we wouldn't have merged the branch.

This commit moves back to checking the ABI of the Platform. However, the
library headers are extracted from the Sdk and injected into the
Platform checkout, so that we can continue filtering private symbols.

Relates to #389