utils/check-abi: Simplify the path to the debuginfo
As it turns out, there is no need to pass the path to the directory
containing the *.debug file for the library being tested. Libabigail is
quite happy to reconstruct this path as long as we give it the path up
to `/usr/lib/debug`.

This makes the code just a little bit simpler.