Rename collect_integration plugin.
Sometimes build fails with the following error if building
specific elements:
Pip package buildstream-external does not contain a plugin named
'collect_integration'

This is due to missing entry point in bst_external for
collect_integration. To be a valid entry point, it has to be renamed
without '-' (dash) character.

It is probably a BuildStream bug that it sometimes work.

Fixes #392.