utils/check-abi: Organize a bit the suppression file
We are eventually going to accumulate a bunch of suppressions in that
file, sometimes multiple suppressions for a given library.

This change moves to the top the libdrm suppressions, and documents
them, so it is obvious at a glance which suppressions are for what.

The 3 suppressions left at the bottom are just temporary things we have
to keep for now, due to either bugs in libabigail or a lack of resources
on our part.