ci: Compare the ABI of a branch to the parent commit
Currently, when someone creates a branch, the CI compares the ABI of
that branch with the ABI of the 18.08 branch, searching for ABI breaks.

This is because 18.08 is the target branch in which the branch being
tested will be merged.

However, this is causing confusion:

1.  contributor creates "mybranch" based on an old 18.08, and pushes it;
2.  the CI compares "mybranch" with the latest 18.08, in which a library
    has been added;
3.  the CI will report an ABI break.

The ABI break disappears as soon as "mybranch" is rebased on the latest
18.08 and pushed again, but requiring that is a barrier for new
contributors.

With this commit, we now compare the ABI of the branch being tested to
the ABI of the parent of the branch, that is, the last commit in 18.08
on top of which the branch is based.

Fixes #402