Add libclc to platform
libclc is a dependency of clinfo, it makes sense to include it in
the platform. This commit also makes us include some static libraries
in llvm6, the ones kept are the minimum required to build libclc,
unfortunately.

libclc uses an SVN repo for development, the git repo is only a mirror,
ideally we would have an SVN plugin for BuildStream. Additionally there
doesn't seem to be any tags for releases in the git repo, likely because
it's a mirror.

I also add a patch to force the custom configure script for libclc to
use python3 by default.