GnuPG: Fix regressions from switching to git
Switching to git caused some regressions that broke builds. In
particular:
* Most elements needed `--enable-maintainer-mode` passed to `configure`
* Some elements needed docs disabling due to a missing dependency on
`fig2dev`.
As this depends on X11 it doesn't seem feasible to include this just for
the docs
* GnuPG ran tests that had data stored in .tar files. Unfortunately
extracting the tar archive required changing the owner of the files,
but this was not possible in the sandbox. Thus this test is disabled by
a patch.