Fix some RPATH generated by libtool
`libtool --mode=link` creates a local binary that has RPATH to unknown
library paths from the runtime at compile time. The known paths are
loaded from /etc/ld.so.conf.

`libtool --mode=install` either copy ("fast-install") that binary or relinks
it. If it relinks, it does not consider the current runtime, but the
one we provide with LT_SYS_LIBRARY_PATH.

Unfortunately, the default mode is "fast-install". We do not want to
disable this mode because it would take more time, and also some
elements fail to compile properly with it (e.g. bintuils).

Other issue is that `/etc/ld.so.conf` is not supposed to be used when
cross-compiling or bootstraping. The current system has nothing to do
with the system we are building.

In order to force libtool to take our configured environment for
libraries, we force the config cache value.