gst-libav: Use 1.15.1
Updates gst-libav to use 1.15.1 rather than the arbitrary git snapshot
used by Debian, as we used to. Two of the patches which we used were
backports which have been merged in the time between the Debian snapshot
and 1.15.1, so are no longer needed. However a patch is added to revert
a commit that used the gstreamer 1.15 API, to remove the dependency.

Note that 1.15.x is still an unstable branch.