linux.bst: unify kernel image installation
Always install the default kernel image. In our supported
architectures, the default is the compressed image.

Before this change, the compressed kernel image was installed for
x86_64 and i686, and the uncompressed image was installed for
ARM and AArch64. Since qemu-system-arm wouldn't boot uncompressed
kernel images, the image produced for this architecture couldn't be
used in QEMU.

This change also ensures that the kernel image is installed using the
the same name in all architectures. Previously the image was named
differently in x86_64/i686 and AArch64/ARM.