Makefile: Add clean-test target, fix test-apps
There are some files created by the `test-apps` target that we don't
clean up, this adds a clean up for them.

This commit also makes various other improvements to the Makefile,
using a variable to invoke `bst --colors $(ARCH_OPTS)` for easier
maintenance and ensuring the default shell is bash.