ladspa-sdk: Force respect for our build flags
The LADSPA SDK makefile overwrites the environment build flags, but also
requires some additional flags, this commit forces the makefile to show
due respect to our configuration.

Somehow the default target for the Makefile is `test` if you change the
default CFLAGS, so this commit also forces the correct target.