components/libglvnd: Update to the latest patch release
We've been using a Git snapshot since November 2018, so we could build
libglvnd with Python 3 instead of Python 2.

Upstream recently released a new version, which only fixes bugs, among
which a race condition and a segmentation fault.

This commit updates to that release, so we both get the bug fixes and
move on from a pre-release snapshot.