project.conf: Make cmake build verbose
CMake is not a very verbose build system by default, and needs the
`CMAKE_VERBOSE_MAKEFILE` flag to be set to ON in order to see what's
going on. In particular, the default generated Makefiles do not tell us
if the build flags defined for our project are being respected by the
package being built. This means some packages may be building without
our hardening flags silently.