bootstrap/symlinks.bst: Adding symlink from /var/run to /run
We prefer to have the run directory in /run rather than /var/run,
but some software still expects things to reside in the standard
/var/run directory

For example, the standard location for the system message bus
is located in /var/run:

  https://dbus.freedesktop.org/doc/dbus-specification.html#message-bus-types-system

As such, the hard coded default for glib to lookup the system bus
(unless environment variables have been carefully specified) will
default to the /var/run location.

Most systems (including debian) include a /var/run -> /run symlink,
to ensure that both paths are synonymous while allowing the run dir
to actually reside at /run, so doing the same here will be the
most frictionless.