Fix self-junctions for cross-compilers
In order to import runtimes with different options we go through junctions
of the project itself. But doing it with the git repository requires us
to track and merge changes in multiple merge requests.

Using a local source with all the files from the root of the project
did not work because it would copy all checkouts into a temporary
directory, which took too long. Selectively import directories and
`project.conf` helps not copying useless files.