gst-plugins-bad: Reenable ladspa
Patch gstreamer-plugins-bad to ignore the optional dependency on LRDF
even if ladspa is explicitly enabled. The current behaviour is that the
optional dependency is required if ladspa is explicitly enabled but
actually optional if plugins are enabled/disabled automagically. Rather
than adding the (at least) two additional libraries, I just patch out
this dependency.