Force cross linker to follow rpath
The linker (ld), when using cross compilation disables following rpath
unless there is a sysroot. In which case it will prepend the sysroot
to the rpath. However since we use multiarch, we know that paths
to libraries are valid across architectures because the path will
contain the platform triple. So we can allow the linker to follow
rpath. To work around the limitation, we tell the linker the
sysroot is /.

Fixes #799.