desktop/mesa.bst: Install a pkgconfig file for glesv2.
The build of new versions of Mesa does not seem to install pkgconfig
file for glesv2 anymore when using libglvnd. It was probably an
oversight. Mesa still installs the headers. Only the library is
provided by libglvnd, and it does not provide a pkgconfig file either.
So for the moment we just generate one manually.