Add an ffmpeg full codecs extension
Adds an extension containing a more featureful ffmpeg (at least the
boiler plate for one, for now). This is intended to be used as the
platform compat extensions.

Additionally alter the gst-libav patch to also not cache codecs found in
`/app/lib/ffmpeg`, the recommended mount point for this extension.