gst-plugins-bad: Use auto features for enablement
We don't want to enable all bad plugins, as some are unmaintained and
may provide security risks. However we don't want to disable them all
either, as there are a *lot* we need to manually turn back on. We should
do that eventually, but we probably don't have time before the 19.08
release.