Remove static libraries when image created
Adds a `static-blacklist` split rule, which may be overwritten on a
per-element basis. Examples of how to do this are seen in the gcc and
glibc elements. Additionally stops handling the static libraries of llvm
in a special way, to simplify structure.

There are a couple of `static-whitelist` rules also defined in gcc and
glibc, but this is never used and is to show explicitly what we're
including.