components/python3.bst: Add dummy pyconfig.h redirecting to correct path
`pyconfig.h` is a platform dependant header that is in a different
path than other headers.  Some projects have very weird ways to find
the include path and do not use `pkg-config` or `python3-config`. They
therefore fail to find `pyconfig.h`. We introduce a dummy `pyconfig.h`
that redirects to the correct path. We also change the order of
includes so that this dummy header is used only as fallback.