components/sdl2.bst: Add dummy SDL_config.h that redirects to correct path
SDL_config.h is architecture dependant. Some software have weird ways
to detect include paths for SDL2 instead of using pkg-config or the
cmake module. Here we add a dummy header redirecting to the correct
path. We also give more priority to the arch-dependent path so that
this dummy header is only used as fallback.