.gitlab-ci.yml: Stop publishing native i686 flatpak runtimes
Some distros doesn't even support this for a while, and very few
hardware is available that only support x86 32 bits only

Also, we want to add SSE2 support for the 32compat extension to
get optimization for steam/wine. Problem is that there are some
very old i686 hardware that do not support SSE2, and the apps
will crash if it tries to run. To solve this we have several options:

1. Remove support for native i686 altogether
2. Keep support for i686 but only for hardware that support SSE2
3. Keep support for all i686 but remove SSE2 optimization
   for the compat32 extension
4. Keep support for all i686 and keep support for SSE2 for the
   compat32 extension (this will require build everything twice)

Taking in account the very little amount of native i686 users
nowadays, and the amount of resources we have, we have decided
to go for 1

Other resources:
- https://fedoraproject.org/wiki/Changes/Stop_Building_i686_Kernels

Fixes #437