Add noopenh264 stub library to platform
Adds the noopenh264 stub library from Endless to the platform, in order
to replace this with the openh264 binary from cisco in an extension.
This will allow users to decode H.264 video without infringing patents.
We also build ffmpeg, gst-plugins-bad with openh264 support now.

I had to patch noopenh264 to update the version and the actual API, but
I opened an upstream MR for this. I also have an MR to update the
openh264 meson.build to actually have the right version.

noopenh264 is placed in `/usr/lib/%{triplet}/openh264/extra` so that it
can be replaced by the actual openh264 binary when an extension is
downloaded. The extra is necessary as this is where flatpak puts things
downloaded as extra data.