Add mountpoint for openh264 platform extension
This allows us to add an openh264 extension in a later MR than this, as
the actual release blocker is adding noopenh264 to the platform.

Our mountpoint is /usr/lib/%{triplet}/openh264, but the openh264 binary
should be placed under /usr/lib/%{triplet}/openh264/extra, as this is
where noopenh264 is configured to be. This is because flatpak defaults
to downloading extra data into the extra/ subdir of an app/extension,
and moving things around makes things slower and unnecessarily complex.

I have an openh264 extension working, but at present there are
discussions on versioning and concerns about installing with a flatpak
which is too old which are blocking these simple changes getting merged.
As a result I am splitting those into a new MR.