Noopenh264: Copy stubs to /usr/lib/%{triple}
It turns out that without loading an extension, these libraries are not
on the search path, and as such cause a breakage when ffmpeg looks for
them. Sorry about this, I really should have tested things work
*without* the extension.

The extension should still work, as anything they specify for the
ld-path takes precedence.