Don't publish ffmpeg-html5 extension
As per the discussion on #854, we are no longer going to publish the
ffmpeg-html5 extension to flathub. If apps want html5 codecs reliably
for now, they should bundle ffmpeg-full while distros update their
flatpak. We will publish the openh264 extension as soon as we figure out
how we're going to do so.