Add an openh264 extension, but not to flathub
As we don't want users to have to keep redownloading things due to
churn, we are not publishing the openh264 extension without further
testing - we need to ensure the ref doesn't change inadvertently.

As openh264 is only supported on flatpak >= 1.4.2, not publishing is not
a big deal, and apps needing H.264 codecs should either bundle their own
or use the ffmpeg-full extension.