Set alsa environment to avoid loading host configuration.
If either `ALSA_CONFIG_DIR` or `ALSA_CONFIG_PATH` are defined on host
system, this causes problems with Flatpak since alsa will try to load
those paths. We now set their values so that they are not picked
by the host.

We also provide an alternative configuration for Flatpak which does
not load "~/.asoundrc" which could be a host configuration in case of
"--filesystem=home". Instead we load `$XDG_CONFIG_HOME/asound.conf`
which is safe because reising on the `.var/app/<id>/config` directory.

Fixes #857