.gitlab-ci.yml: Publish openh264 ext to flathub
As we have confirmed that installing with old flatpaks is not
catastrophic, and that even if there is churn it's not a huge deal, I
think it's safe to publish this now.

At present this should be of use to users of arch and its derivatives,
fedora >=30, endless and presumably more distros I'm not sure on.

Time for royalty free H.264 decoding!