Create intermediate manifest for build dependencies of filters
Filters use build dependencies to the real element. The source which
is used to infer CPE data is taken from that element and would not end
up in the final manifest. This adds runtime dependencies to extracted
manifest for those sources so that we do not have to replicate
manually version information on the filters.