Fix redundent entries in generated manifest
Some elements have a `-base.bst` file and then multiple filter
elements to split the element. Because the source is on the
`-base.bst` which filter elements depends only for build, the source
is not added automatically in the manifest. To solve this we use a
`-manifest.bst` manifest element. Unfortunately if the `-base.bst` has
runtime dependencies, these will be pulled from that manifest element.
So `-base.bst` should only have build dependencies.