create mirror_management scripts
created check_mirrors.py, a script that reads in a url manifest (as
generated by the url_manifest plugin), and runs tests to check that
everything is mirrored as is should be. There are 5 tests, each is
optional, and they can be run in whatever combination is needed, using
command-line options.

Tests that: aliases are used in source urls, that aliases have
download-mirrors defined, that the mirrors actually exist, and (for git
and ostree sources) that the appropriate commit/ref exists in the source
and mirror respositories.

Also has an ignore_file, to whitelist certain elements or source urls so
that they won't be reported as errors. Sources can be ignored by all
tests, or only for specific tests.