add url_manifest plugin
the new version of the url_manifest plugin lists only the main source URLs, not
the urls of our download mirrors. Accessing the list of mirror urls / mirror
aliases was hard to do from inside the plugin. Instead we'll write a script to
modify the manifest after it's generated. (To come in a future commit.)

added a new alias for crates, so we can get urls for rust cargo crate sources.