Add mirror-checking scripts to gitlab CI
Check_mirrors.py will run in the gitlab CI, when new commits are pushed,
and on the nightly schedules.

When new commits are pushed, only the alias tests will run: checking aliases
are used in urls, and that each one is defined in the mirrors yaml file.
(These two tests are looking for issues in the freedesktop-sdk
repository itself, and detect issues that develpers can (and should)
fix themselves)

In the nightly scheduled job, all tests run. If unmirrored sources are
detected, then the job artefact will contain the scripts and information
needed to create new mirrors. (When these tests fail, it can mean that
mirrors need to be created or modified, and only certain people have
permissions to resolve these issues.)