handle failures in scan-for-dev-files.sh
tests/check-dev-files.bst inovokes scan-for-dev-files.sh, and the build
is supposed to fail if scan-for-dev-files.sh fails. But failures were
not being recognized, because of a pipe used in line 10.

rewrite check-dev-files to avoid using the pipe.

also condense the find commands in scan-for-dev files, into a single
command. Test should now run faster. Script will no longer fail if
"${1}/usr/include" doesn't exist.